Move models into their own package
