--- models.go
@@ -1,11 +1,0 @@
-package main
-
-type AirportDetails struct {
-	City     string
-	FullName string
-}
-
-type RouteDetails struct {
-	Destination         string
-	TravelTimeInMinutes int32
-}
