+++ route_list_airport.go
@@ -3,16 +3,18 @@ package main
 import (
 	"fmt"
 	"net/http"
+
+	"github.com/tombuildsstuff/planes/models"
 )
 
 // listAirports handles `/airports`
 func listAirports(w http.ResponseWriter, r *http.Request) {
-	obj := ListAirportsResponse{
-		Airports: func() []Airport {
-			out := make([]Airport, 0)
+	obj := models.ListAirportsResponse{
+		Airports: func() []models.Airport {
+			out := make([]models.Airport, 0)
 
 			for code, airport := range airports {
-				out = append(out, Airport{
+				out = append(out, models.Airport{
 					Name: airport.FullName,
 					Code: code,
 					Uri:  fmt.Sprintf("/airports/%s", code),
@@ -23,13 +25,4 @@ func listAirports(w http.ResponseWriter, r *http.Request) {
 		}(),
 	}
 	serializeJson(obj, w)
-}
-
-type ListAirportsResponse struct {
-	Airports []Airport `json:"airports"`
-}
-type Airport struct {
-	Name string `json:"name"`
-	Code string `json:"code"`
-	Uri  string `json:"uri"`
 }
