+++ route_list_routes.go
@@ -3,32 +3,24 @@ package main
 import (
 	"fmt"
 	"net/http"
+
+	"github.com/tombuildsstuff/planes/models"
 )
 
 // listRoutes handles /routes
 func listRoutes(w http.ResponseWriter, r *http.Request) {
-	out := make([]RouteSummary, 0)
+	out := make([]models.RouteSummary, 0)
 	for originCode, destinations := range routes {
 		for _, destination := range destinations {
-			out = append(out, RouteSummary{
+			out = append(out, models.RouteSummary{
 				OriginCode:      originCode,
 				DestinationCode: destination.Destination,
 				Uri:             fmt.Sprintf("/route/%s/%s", originCode, destination.Destination),
 			})
 		}
 	}
-	obj := ListRoutesResponse{
+	obj := models.ListRoutesResponse{
 		Routes: out,
 	}
 	serializeJson(obj, w)
-}
-
-type ListRoutesResponse struct {
-	Routes []RouteSummary `json:"routes"`
-}
-
-type RouteSummary struct {
-	OriginCode      string `json:"origin_code"`
-	DestinationCode string `json:"destination_code"`
-	Uri             string `json:"uri"`
 }
